requirements.txt
python-chess
pandas
click
scikit-learn
s3fs
plotly
requests
pytest
coverage
black
tqdm==4.48.2
